ticket-collector/requirements.txt
prometheus_client==0.6.0
APScheduler==3.6.0
-f https://s3pypi.mettel-automation.net/igzpackages/index.html
igzpackages==2.0.1
pymongo==3.11.4
dependency-injector==4.4.0
python-dotenv==0.17.1
requests==2.25.1
